backend/requirements.txt
absl-py==2.2.2
astunparse==1.6.3
bidict==0.23.1
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
dnspython==2.7.0
eventlet==0.40.0
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.0
Flask-JWT-Extended==4.7.1
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
fsspec==2025.5.0
gast==0.6.0
google-pasta==0.2.0
greenlet==3.2.2
grpcio==1.71.0
h11==0.16.0
h5py==3.13.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
keras==3.10.0
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.0.9
networkx==3.4.2
numpy==2.2.6
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
pillow==11.2.1
protobuf==6.31.0
Pygments==2.19.1
PyJWT==2.10.1
pyotp==2.9.0
python-engineio==4.12.1
python-socketio==5.13.0
pytz==2025.2
qrcode==8.2
requests==2.32.3
rich==14.0.0
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.41
SQLAlchemy-Utils==0.41.2
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
torch==2.7.0
torchvision==0.22.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
wsproto==1.2.0
